feat: 🎸 Added a handler to manage state
