+++ libs/extract/src/lib/dropdown/dropdown.spec.ts
@@ -1,401 +1,256 @@
-import { createPopper } from '@popperjs/core'
-import {
-  activate,
-  close,
-  create,
-  deactivate,
-  observe,
-  open,
-  select,
-  toggle,
-  unobserve,
-} from './dropdown'
-import { AbstractDropdown, DropdownOption } from './types'
+import { createPopper, Instance } from '@popperjs/core'
+import { AbstractDropdown, DropdownHandler, DropdownListener, DropdownOption } from "./types"
+import { create } from './dropdown'
 
 jest.mock('@popperjs/core', () => ({
   createPopper: jest.fn().mockName('createPopper')
 }))
 
+const tick = (t = 0) => new Promise((r) => setTimeout(r, t))
+
 describe('dropdown', () => {
-  let id: string
-  let options: DropdownOption[]
-  beforeEach(() => {
-    id = 'foo'
-    options = [
+  let handler: DropdownHandler
+  let listener: jest.Mock<DropdownListener>
+
+  let toggler: HTMLElement
+  let listbox: HTMLElement
+
+  let popper: Partial<Instance>
+
+  beforeEach(async () => {
+    toggler = document.createElement('button')
+    listbox = document.createElement('div')
+    listener = jest.fn<never, [AbstractDropdown]>().mockName('listener')
+    const options: DropdownOption[] = [
       { key: 'A', value: 1 },
       { key: 'B', value: 2 },
+      { key: 'C', value: 3 },
     ]
+    popper = {
+      update: jest.fn()
+        .mockName('update')
+        .mockResolvedValue({ styles: { popper: {} } }),
+      destroy: jest.fn().mockName('destroy')
+    }
+    ;(createPopper as jest.Mock<Instance>)
+      .mockReturnValue(popper as Instance)
+    
+    handler = create({ id: 'foo', options }, toggler, listbox, listener)
+    await tick()
   })
-  describe('create', () => {
-    it('uses passed in id', () => {
-      const dropdown = create({ id, options })
 
-      expect(dropdown.id).toEqual('foo')
+  afterEach(() => {
+    toggler.remove()
+    listbox.remove()
+    handler.destroy()
+    ;(createPopper as jest.Mock<Instance>).mockReset()
+  })
+
+  describe('methods', () => {
+    it('calls back with a new instance', () => {
+      const initial = handler.dropdown
+      handler.open()
+      expect(handler.dropdown).not.toBe(initial)
     })
-    it('creates an id if not supplied', () => {
-      const dropdown = create({ options })
-
-      expect(dropdown.id).toBeTruthy()
+    it('calls back on activate', async () => {
+      await handler.active(true)
+      expect(listener).toHaveBeenCalledWith(handler.dropdown)
     })
-    describe('options', () => {
-      it('contains options', () => {
-        const dropdown = create({ options })
-  
-        expect(dropdown.options).toHaveLength(2)
-      })
-      it('sets option ids', () => {
-        const dropdown = create({ id, options })
-        const optionAttributes = dropdown.options.map((o) => o.attributes)
-
-        expect(optionAttributes.map((o) => o.id)).toEqual([
-          'foo_option0', 'foo_option1',
-        ])
-      })
-      it('sets option roles', () => {
-        const dropdown = create({ id, options })
-        const optionAttributes = dropdown.options.map((o) => o.attributes)
-
-        expect(optionAttributes.map((o) => o.role)).toEqual([
-          'option', 'option',
-        ])
-      })
-      it('sets option selected', () => {
-        options[1].selected = true
-        const dropdown = create({ id, options })
-        const optionAttributes = dropdown.options.map((o) => o.attributes)
-
-        expect(optionAttributes.map((o) => o['aria-selected'])).toEqual([
-          undefined, true,
-        ])
-      })
+    it('calls back on deactivate', async () => {
+      await handler.active(false)
+      expect(listener).toHaveBeenCalledWith(handler.dropdown)
     })
-    describe('elements', () => {
-      it('sets correct toggle attributes', () => {
-        const dropdown = create({ id, options })
-
-        expect(dropdown.elements.toggler.attributes).toEqual({
-          'aria-haspopup': 'listbox',
-          'aria-owns': 'foo',
-          'aria-expanded': false,
-        })
-      })
-      it('sets correct toggler classes', () => {
-        const dropdown = create({ id, options })
-
-        expect(dropdown.elements.toggler.classes).toEqual(['dropdown-toggle'])
-      })
-      it('sets correct listbox attributes', () => {
-        const dropdown = create({ id, options })
-
-        expect(dropdown.elements.listbox.attributes).toEqual({
-          id,
-          role: 'listbox',
-          tabIndex: -1,
-        })
-      })
-      it('sets correct listbox classes', () => {
-        const dropdown = create({ id, options })
-
-        expect(dropdown.elements.listbox.classes).toEqual(['popover'])
-      })
+    it('calls back on open', async () => {
+      await handler.open()
+      expect(listener).toHaveBeenCalledWith(handler.dropdown)
+    })
+    it('calls back on close', async () => {
+      await handler.close()
+      expect(listener).toHaveBeenCalledWith(handler.dropdown)
+    })
+    it('calls back on toggle', async () => {
+      await handler.toggle()
+      expect(listener).toHaveBeenCalledWith(handler.dropdown)
+    })
+    it('calls back on select', async () => {
+      await handler.select(1)
+      expect(listener).toHaveBeenCalledWith(handler.dropdown)
     })
   })
-  describe('interactions', () => {
-    let dropdown: AbstractDropdown
-    beforeEach(() => {
-      dropdown = create({ id, options })
-    })
-    it('defaults to closed', () => {
-      expect(dropdown.isOpen).toBe(false)
-    })
-    it('defaults to inactive', () => {
-      expect(dropdown.isActive).toBe(false)
-    })
-    describe('open', () => {
-      beforeEach(() => {
-        dropdown = open(dropdown)
-      })
-      it('sets isOpen=true', () => {
-        expect(dropdown.isOpen).toBe(true)
-      })
-      it('sets toggler.aria-expanded=true', () => {
-        const expanded = dropdown.elements.toggler.attributes['aria-expanded']
-        expect(expanded).toBe(true)
-      })
-      it('adds listbox class active', () => {
-        const { classes } = dropdown.elements.listbox
-        expect(classes).toContain('active')
-      })
-      describe('close', () => {
-        beforeEach(() => {
-          dropdown = close(dropdown)
-        })
-        it('sets isOpen=false', () => {
-          expect(dropdown.isOpen).toBe(false)
-        })
-        it('sets toggler.aria-expanded=false', () => {
-          const expanded = dropdown.elements.toggler.attributes['aria-expanded']
-          expect(expanded).toBe(false)
-        })
-        it('removes listbox class active', () => {
-          const { classes } = dropdown.elements.listbox
-          expect(classes).not.toContain('active')
-        })
-      })
-    })
-    describe('activate', () => {
-      beforeEach(() => {
-        dropdown = activate(dropdown)
-      })
-      it('sets isActive to true', () => {
-        expect(dropdown.isActive).toBe(true)
-      })
-      describe('deactivate', () => {
-        beforeEach(() => {
-          dropdown = deactivate(dropdown)
-        })
-        it('sets isActive to false', () => {
-          expect(dropdown.isActive).toBe(false)
-        })
-      })
-    })
-    describe('toggle', () => {
-      it('switches between open and close', () => {
-        expect(dropdown.isOpen).toBe(false)
-        dropdown = toggle(dropdown)
-        expect(dropdown.isOpen).toBe(true)
-        dropdown = toggle(dropdown)
-        expect(dropdown.isOpen).toBe(false)
-      })
-    })
-    describe('select', () => {
-      beforeEach(() => {
-        dropdown = open(dropdown)
-      })
-      it('sets option selected', () => {
-        dropdown = select(dropdown, dropdown.options[1])
-        const optionAttributes = dropdown.options.map((o) => o.attributes)
 
-        expect(optionAttributes.map((o) => o['aria-selected'])).toEqual([
-          undefined, true,
-        ])
-      })
-      it('sets activedecendant', () => {
-        const selectedOption = dropdown.options[1]
-        dropdown = select(dropdown, selectedOption)
-        const attrs = dropdown.elements.listbox.attributes
+  describe('keyboard', () => {
+    describe('when inactive', () => {
+      it('does not trigger key events', async () => {
+        document.dispatchEvent(new KeyboardEvent('keydown', { key: 'ArrowDown' }))
+        document.dispatchEvent(new KeyboardEvent('keydown', { key: 'ArrowUp' }))
+        document.dispatchEvent(new KeyboardEvent('keydown', { key: 'Escape' }))
+        document.dispatchEvent(new KeyboardEvent('keydown', { key: 'Home' }))
+        document.dispatchEvent(new KeyboardEvent('keydown', { key: 'End' }))
+        document.dispatchEvent(new KeyboardEvent('keydown', { key: ' ' }))
+        await tick()
 
-        expect(attrs['aria-activedescendant']).toEqual(selectedOption.attributes.id)
-      })
-      it('does not close dropdown', () => {
-        dropdown = select(dropdown, dropdown.options[1])
-        expect(dropdown.isOpen).toBe(true)
-      })
-      describe('with step', () => {
-        it('sets first if no selection and step is 1', () => {
-          dropdown = select(dropdown, 1)
-          
-          expect(dropdown.options[0].selected).toBe(true)
-        })
-        it('sets first if no selection and step is -1', () => {
-          dropdown = select(dropdown, -1)
-          
-          expect(dropdown.options[0].selected).toBe(true)
-        })
-        it('sets second if first is selected and step is 1', () => {
-          dropdown.options[0].selected = true
-          dropdown = select(dropdown, 1)
-          
-          expect(dropdown.options[1].selected).toBe(true)
-        })
-        it('stays if last is selected and step is 1', () => {
-          dropdown.options[1].selected = true
-          dropdown = select(dropdown, 1)
-          
-          expect(dropdown.options[1].selected).toBe(true)
-        })
-        it('sets second if first is selected and step is 1', () => {
-          dropdown.options[1].selected = true
-          dropdown = select(dropdown, -1)
-          
-          expect(dropdown.options[0].selected).toBe(true)
-        })
-        it('stays if first is selected and step is -1', () => {
-          dropdown.options[0].selected = true
-          dropdown = select(dropdown, -1)
-          
-          expect(dropdown.options[0].selected).toBe(true)
-        })
-        describe('loop', () => {
-          beforeEach(() => {
-            dropdown = activate({ ...dropdown, loop: true })
-          })
-          it('sets last if no selection and step is -1', () => {
-            dropdown = select(dropdown, -1)
-            
-            expect(dropdown.options[1].selected).toBe(true)
-          })
-          it('sets first if last is selected and step is 1', () => {
-            dropdown.options[1].selected = true
-            dropdown = select(dropdown, 1)
-            
-            expect(dropdown.options[0].selected).toBe(true)
-          })
-          it('sets last if first is selected and step is -1', () => {
-            dropdown.options[0].selected = true
-            dropdown = select(dropdown, -1)
-            
-            expect(dropdown.options[1].selected).toBe(true)
-          })
-        })
+        expect(listener).toHaveBeenCalledTimes(1)
       })
     })
-    describe('observe', () => {
-      let listener: jest.Mock<undefined, [AbstractDropdown]>
-      beforeEach(() => {
-        listener = jest.fn()
-        observe(dropdown, listener)
+    describe('when active', () => {
+      beforeEach(async () => {
+        await handler.active(true)
       })
-      afterEach(() => unobserve(dropdown))
-      describe('when inactive', () => {
-        it('does not trigger key events', () => {
-          document.dispatchEvent(new KeyboardEvent('keydown', { key: 'ArrowDown' }))
-          document.dispatchEvent(new KeyboardEvent('keydown', { key: 'ArrowUp' }))
-          document.dispatchEvent(new KeyboardEvent('keydown', { key: 'Escape' }))
-          document.dispatchEvent(new KeyboardEvent('keydown', { key: 'Home' }))
-          document.dispatchEvent(new KeyboardEvent('keydown', { key: 'End' }))
+      afterEach(() => handler.active(false))
+      describe('Space', () => {
+        it('opens closed dropdown', async () => {
           document.dispatchEvent(new KeyboardEvent('keydown', { key: ' ' }))
+          await tick(10)
 
-          expect(listener).not.toHaveBeenCalled()
+          expect(listener).toHaveBeenCalledTimes(3)
+          const [dd] = listener.mock.calls[2]
+          expect(dd.isOpen).toBe(true)
+        })
+        it('closes open dropdown', async () => {
+          document.dispatchEvent(new KeyboardEvent('keydown', { key: ' ' }))
+          document.dispatchEvent(new KeyboardEvent('keydown', { key: ' ' }))
+          await tick()
+
+          expect(listener).toHaveBeenCalledTimes(4)
+          const [dd] = listener.mock.calls[3]
+          expect(dd.isOpen).toBe(false)
         })
       })
-      describe('when active', () => {
-        beforeEach(() => {
-          dropdown.options = dropdown.options.map((o) => ({ ...o, selected: false }))
-          dropdown = activate(close(dropdown))
-          observe(dropdown, listener)
-        })
-        describe('Space', () => {
-          it('opens closed dropdown', () => {
-            document.dispatchEvent(new KeyboardEvent('keydown', { key: ' ' }))
-            
-            expect(listener).toHaveBeenCalled()
-            const [dd] = listener.mock.calls[0]
-            expect(dd.isOpen).toBe(true)
-          })
-          it('closes open dropdown', () => {
-            dropdown = observe(open(dropdown), listener)
-            document.dispatchEvent(new KeyboardEvent('keydown', { key: ' ' }))
-            
-            expect(listener).toHaveBeenCalled()
-            const [dd] = listener.mock.calls[0]
-            expect(dd.isOpen).toBe(false)
-          })
-        })
-        describe('Escape', () => {
-          it('closes open dropdown', () => {
-            dropdown = observe(open(dropdown), listener)
-            document.dispatchEvent(new KeyboardEvent('keydown', { key: 'Escape' }))
+      describe('Escape', () => {
+        it('closes open dropdown', async () => {
+          document.dispatchEvent(new KeyboardEvent('keydown', { key: ' ' }))
+          document.dispatchEvent(new KeyboardEvent('keydown', { key: 'Escape' }))
+          await tick()
 
-            expect(listener).toHaveBeenCalled()
-            const [dd] = listener.mock.calls[0]
-            expect(dd.isOpen).toBe(false)
-          })
-          it('does nothing to closed dropdown', () => {
-            document.dispatchEvent(new KeyboardEvent('keydown', { key: 'Escape' }))
+          expect(listener).toHaveBeenCalledTimes(4)
+          const [dd] = listener.mock.calls[3]
+          expect(dd.isOpen).toBe(false)
+        })
+        it('does nothing to closed dropdown', async () => {
+          expect(listener).toHaveBeenCalledTimes(2)
+          document.dispatchEvent(new KeyboardEvent('keydown', { key: 'Escape' }))
+          await tick()
 
-            expect(listener).not.toHaveBeenCalled()
-          })
-        })
-        describe('Arrow down', () => {
-          it('opens closed dropdown', () => {
-            document.dispatchEvent(new KeyboardEvent('keydown', { key: 'ArrowDown' }))
-  
-            expect(listener).toHaveBeenCalled()
-            const [dd] = listener.mock.calls[0]
-            expect(dd.isOpen).toBe(true)
-          })
-          it('selects first option', () => {
-            dropdown = select(dropdown, dropdown.options[0])
-            document.dispatchEvent(new KeyboardEvent('keydown', { key: 'ArrowDown' }))
-  
-            expect(listener).toHaveBeenCalled()
-            const [dd] = listener.mock.calls[0]
-            expect(dd.options[0].selected).toBe(true)
-          })
-        })
-        describe('Arrow up', () => {
-          it('opens closed dropdown', () => {
-            document.dispatchEvent(new KeyboardEvent('keydown', { key: 'ArrowUp' }))
-  
-            expect(listener).toHaveBeenCalled()
-            const [dd] = listener.mock.calls[0]
-            expect(dd.isOpen).toBe(true)
-          })
-          it('selects first option', () => {
-            dropdown = select(dropdown, dropdown.options[0])
-            document.dispatchEvent(new KeyboardEvent('keydown', { key: 'ArrowUp' }))
-  
-            expect(listener).toHaveBeenCalled()
-            const [dd] = listener.mock.calls[0]
-            expect(dd.options[0].selected).toBe(true)
-          })
-          it('selects last option if looped', () => {
-            dropdown.loop = true
-            document.dispatchEvent(new KeyboardEvent('keydown', { key: 'ArrowUp' }))
-  
-            expect(listener).toHaveBeenCalled()
-            const [dd] = listener.mock.calls[0]
-            expect(dd.options[1].selected).toBe(true)
-          })
-        })
-        describe('Home', () => {
-          it('opens closed dropdown', () => {
-            document.dispatchEvent(new KeyboardEvent('keydown', { key: 'Home' }))
-  
-            expect(listener).toHaveBeenCalled()
-            const [dd] = listener.mock.calls[0]
-            expect(dd.isOpen).toBe(true)
-          })
-          it('selects first option', () => {
-            dropdown = select(dropdown, dropdown.options[1])
-            document.dispatchEvent(new KeyboardEvent('keydown', { key: 'Home' }))
-  
-            expect(listener).toHaveBeenCalled()
-            const [dd] = listener.mock.calls[0]
-            expect(dd.options[0].selected).toBe(true)
-          })
-        })
-        describe('End', () => {
-          it('opens closed dropdown', () => {
-            document.dispatchEvent(new KeyboardEvent('keydown', { key: 'End' }))
-  
-            expect(listener).toHaveBeenCalled()
-            const [dd] = listener.mock.calls[0]
-            expect(dd.isOpen).toBe(true)
-          })
-          it('selects last option', () => {
-            document.dispatchEvent(new KeyboardEvent('keydown', { key: 'End' }))
-  
-            expect(listener).toHaveBeenCalled()
-            const [dd] = listener.mock.calls[0]
-            expect(dd.options[1].selected).toBe(true)
-          })
+          expect(listener).toHaveBeenCalledTimes(2)
         })
       })
-      describe('popper', () => {
-        beforeEach(() => {
-          global.innerWidth = 1024
-          dropdown = activate(close(dropdown))
-          observe(dropdown, listener)
+      describe('Arrow down', () => {
+        it('opens closed dropdown', async () => {
+          document.dispatchEvent(new KeyboardEvent('keydown', { key: 'ArrowDown' }))
+          await tick()
+
+          expect(listener).toHaveBeenCalledTimes(3)
+          const [dd] = listener.mock.calls[2]
+          expect(dd.isOpen).toBe(true)
         })
-        it('calls createPopper if width < 576', () => {
-          global.innerWidth = 500
-          global.dispatchEvent(new UIEvent('resize'))
-          
-          expect(createPopper).toHaveBeenCalled()
+        it('selects first option', async () => {
+          document.dispatchEvent(new KeyboardEvent('keydown', { key: 'ArrowDown' }))
+          await tick()
+
+          expect(listener).toHaveBeenCalledTimes(3)
+          const [dd] = listener.mock.calls[2]
+          expect(dd.options[0].selected).toBe(true)
         })
       })
+      describe('Arrow up', () => {
+        it('opens closed dropdown', async () => {
+          document.dispatchEvent(new KeyboardEvent('keydown', { key: 'ArrowUp' }))
+          await tick()
+
+          expect(listener).toHaveBeenCalledTimes(3)
+          const [dd] = listener.mock.calls[2]
+          expect(dd.isOpen).toBe(true)
+        })
+        it('selects first option', async () => {
+          document.dispatchEvent(new KeyboardEvent('keydown', { key: 'ArrowUp' }))
+          await tick()
+
+          expect(listener).toHaveBeenCalledTimes(3)
+          const [dd] = listener.mock.calls[2]
+          expect(dd.options[0].selected).toBe(true)
+        })
+        it('selects last option if looped', async () => {
+          handler.loop(true)
+          document.dispatchEvent(new KeyboardEvent('keydown', { key: 'ArrowUp' }))
+          await tick()
+
+          expect(listener).toHaveBeenCalledTimes(4)
+          const [dd] = listener.mock.calls[3]
+          expect(dd.options[2].selected).toBe(true)
+        })
+      })
+      describe('Home', () => {
+        it('opens closed dropdown', async () => {
+          document.dispatchEvent(new KeyboardEvent('keydown', { key: 'Home' }))
+          await tick()
+
+          expect(listener).toHaveBeenCalledTimes(3)
+          const [dd] = listener.mock.calls[2]
+          expect(dd.isOpen).toBe(true)
+        })
+        it('selects first option', async () => {
+          document.dispatchEvent(new KeyboardEvent('keydown', { key: 'Home' }))
+          await tick()
+
+          expect(listener).toHaveBeenCalledTimes(3)
+          const [dd] = listener.mock.calls[2]
+          expect(dd.options[0].selected).toBe(true)
+        })
+      })
+      describe('End', () => {
+        it('opens closed dropdown', async () => {
+          document.dispatchEvent(new KeyboardEvent('keydown', { key: 'End' }))
+          await tick()
+
+          expect(listener).toHaveBeenCalledTimes(3)
+          const [dd] = listener.mock.calls[2]
+          expect(dd.isOpen).toBe(true)
+        })
+        it('selects last option', async () => {
+          document.dispatchEvent(new KeyboardEvent('keydown', { key: 'End' }))
+          await tick()
+
+          expect(listener).toHaveBeenCalledTimes(3)
+          const [dd] = listener.mock.calls[2]
+          expect(dd.options[2].selected).toBe(true)
+        })
+      })
+    })
+  })
+  describe('popper', () => {
+    beforeEach(() => {
+      global.innerWidth = 1024
+    })
+    afterEach(() => {
+      global.innerWidth = 1024
+    })
+    it('calls createPopper (since standard width is 1024)', () => {
+      expect(createPopper).toHaveBeenCalledTimes(1)
+    })
+    it('removes popper if innerWidth < 576', async () => {
+      global.innerWidth = 500
+      global.dispatchEvent(new UIEvent('resize'))
+      await tick()
+
+      expect(listener).toHaveBeenCalledTimes(2)
+      expect(popper.destroy).toHaveBeenCalledTimes(1)
+      expect(handler.popper).toBe(undefined)
+    })
+    it('reattaches popper if innerWidth >= 576', async () => {
+      global.innerWidth = 500
+      global.dispatchEvent(new UIEvent('resize'))
+      await tick()
+      expect(listener).toHaveBeenCalledTimes(2)
+      expect(popper.destroy).toHaveBeenCalledTimes(1)
+      expect(handler.popper).toBe(undefined)
+
+      global.innerWidth = 576
+      global.dispatchEvent(new UIEvent('resize'))
+      await tick(100)
+
+      expect(listener).toHaveBeenCalledTimes(3)
+      expect(createPopper).toHaveBeenCalledTimes(2)
+      expect(handler.popper).toBe(popper)
     })
   })
 })
