+++ libs/extract/src/lib/dropdown/index.ts
@@ -1,4 +1,4 @@
-export * from './dropdown'
+export * from './reducers'
 
 export {
   AbstractDropdown,
