+++ libs/extract/src/lib/dropdown/types.ts
@@ -1,5 +1,6 @@
 import { Observable, Subscription } from 'rxjs'
 import { AriaAttributes } from 'react'
+import { Instance } from '@popperjs/core'
 
 export interface DropdownOption {
   key: string
@@ -35,4 +36,22 @@ export interface AbstractDropdown {
   }>
   keyboardEvents: Observable<KeyboardEvent>
   subscription: Subscription
+}
+
+export type DropdownListener = (dropdown: AbstractDropdown) => void
+
+export interface DropdownHandler {
+  dropdown: AbstractDropdown
+  toggler: HTMLElement
+  listbox: HTMLElement
+  popper?: Instance
+  subscription: Subscription
+
+  active: (isActive: boolean) => Promise<void>
+  loop: (isLooping: boolean) => Promise<void>
+  open: () => Promise<void>
+  close: () => Promise<void>
+  toggle: () => Promise<void>
+  select: (selection: ExtendedDropdownOption | number) => Promise<void>
+  destroy: () => void
 }
